Attempted fixes for YouTube viewer Manu's bug rpt.
Hi Manu.  Thank you for watching the video and using the script.

Although I cannot reproduce your error, I believe it is happening and tried some guesses to resolve it.  
Go ahead and download the newer code and let me know if it resolves the issue (if you are experiencing it, chances are others are too).

Latest code is checked into github here: 

Items I more defensively coded to try and prevent the error:
1. Found and fixed one CASE...WHEN...END by adding an ELSE clause (probably not the root cause)
2. Added nine CAST(x as varchar(10)) statements around numeric data type sizes in case they aren't auto converting to text like they are for me.
3. In case data_type came back null from query, wrapped it in COALESCE(x, 'Unknown') to prevent concatenating a null